backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.30
alembic==1.13.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.4
PyMySQL==1.1.1
cryptography==42.0.8
jinja2==3.1.4
reportlab==4.2.2
pandas==2.2.2
requests==2.31.0
Faker==25.2.0
